microservices/generator/requirements.txt
kafka-python==2.2.6
pandas
flask